Major update adding some ECDH primitives:   - EC CDH (Elliptic Curve Cryptography Cofactor Diffie-Hellman) as defined in   NIST SP800-56A Section 5.7.1.2.   - X25519 and X448 functions and Diffie-Hellman primitives as defined in   RFC 7748.
Also, various minor fixes have been added to adapt to the new ECDH primitives
integration in libecc framework.

+++ src/curves/aff_pt.c
@@ -101,7 +101,7 @@ int is_on_shortw_curve(fp_src_t x, fp_src_t y, ec_shortw_crv_src_t curve, int *o
 {
 	fp tmp1, tmp2;
 	int ret, cmp;
-	tmp1.magic = tmp2.magic = 0;
+	tmp1.magic = tmp2.magic = WORD(0);
 
 	ret = ec_shortw_crv_check_initialized(curve); EG(ret, err);
 	ret = fp_check_initialized(x);  EG(ret, err);
