+++ src/hash/sm3.c
@@ -309,7 +309,7 @@ int sm3_final(sm3_context *ctx, u8 output[SM3_DIGEST_SIZE])
 	PUT_UINT32_BE(ctx->sm3_state[7], output, 28);
 
 	/* Tell that we are uninitialized */
-	ctx->magic = 0;
+	ctx->magic = WORD(0);
 
 	ret = 0;
